US510647-Refactor-Logout-Implementatiom - not clean if the error callback is called because of the locksesion feature.

+++ mas-foundation/src/main/java/com/ca/mas/foundation/MASUser.java
@@ -415,7 +415,8 @@ public abstract class MASUser implements MASMessenger, MASUserIdentity, ScimUser
                     @Override
                     public void onError(Throwable e) {
                         // - Paramenter to delete or not the local storage
-                        if (force) {
+                        boolean lockSession = getCurrentUser() != null ? getCurrentUser().isSessionLocked() : false;
+                        if (force && !lockSession) {
                             try {
                                 tokenManager.deleteIdToken();
                                 tokenManager.deleteSecureIdToken();
